chore(refactor): #1160 created projects module and moved main component to app component
