+++ web/src/app/app-routing.module.ts
@@ -5,11 +5,15 @@ const routes: Routes = [
     {
         path: '',
         loadChildren: './main/main.module#MainModule'
-      }
+    },
+    {
+        path: 'projects',
+        loadChildren: './projects/projects.module#ProjectsModule'
+    }
 ];
 
 @NgModule({
     imports: [RouterModule.forRoot(routes)],
     exports: [RouterModule]
 })
-export class AppRoutingModule {}
+export class AppRoutingModule { }
