+++ web/src/app/app.component.html
@@ -1,0 +1,78 @@
+<mat-sidenav-container class="sidenav-container">
+    <mat-sidenav #drawer class="sidenav" fixedInViewport="true" role="navigation" mode="over">
+        <div class="heading" fxLayout="column" fxLayoutGap="5px">
+            <img *ngIf="isAuthenticated()" src="{{ getProfile().avatarUrl }}" />
+            <mat-icon *ngIf="!isAuthenticated()">person</mat-icon>
+            <div class="mat-subheading-2">{{ isAuthenticated() ? getProfile().name : 'Guest' }}</div>
+            <div class="mat-caption">{{ isAuthenticated() ? getProfile().email : 'Not logged in' }}</div>
+        </div>
+
+        <mat-nav-list>
+            <ng-container *ngFor="let item of publicRoutes">
+                <a mat-list-item [routerLink]="[item.route]" (click)="drawer.close()">
+                    <mat-icon>{{ item.icon }}</mat-icon>
+                    {{ item.title }}
+                </a>
+            </ng-container>
+
+            <mat-divider></mat-divider>
+
+            <ng-container *ngIf="isAuthenticated()">
+                <ng-container *ngFor="let item of privateRoutes">
+                    <a mat-list-item [routerLink]="[item.route]" (click)="drawer.close()">
+                        <mat-icon>{{ item.icon }}</mat-icon>
+                        {{ item.title }}
+                    </a>
+                </ng-container>
+            </ng-container>
+
+            <mat-divider></mat-divider>
+
+            <ng-container *ngFor="let item of generalRoutes">
+                <a mat-list-item [routerLink]="[item.route]" (click)="drawer.close()">
+                    <mat-icon>{{ item.icon }}</mat-icon>
+                    {{ item.title }}
+                </a>
+            </ng-container>
+
+            <a mat-list-item (click)="login(); drawer.close()" *ngIf="!isAuthenticated()" dashboardLogin>
+                <mat-icon>vpn_key</mat-icon> Login
+            </a>
+            <a mat-list-item (click)="logout(); drawer.close()" *ngIf="isAuthenticated()">
+                <mat-icon>exit_to_app</mat-icon> Sign out
+            </a>
+        </mat-nav-list>
+    </mat-sidenav>
+
+    <mat-sidenav-content>
+        <mat-toolbar class="app-toolbar mat-elevation-z2" fxLayout="row" fxLayoutGap="10px" color="accent">
+            <button mat-icon-button type="button" aria-label="Toggle sidenav" (click)="drawer.toggle()">
+                <mat-icon aria-label="Side nav toggle icon">menu</mat-icon>
+            </button>
+            <span>PipelineDashboard</span>
+            <span fxFlex></span>
+            <span *ngIf="!isAuthenticated()" fxHide.xs>Welcome Guest</span>
+            <span *ngIf="isAuthenticated()" fxHide.xs>{{ getProfile().name }}</span>
+            <button mat-icon-button matTooltip="Login" (click)="login()" *ngIf="!isAuthenticated()" dashboardLogin>
+                <mat-icon>vpn_key</mat-icon>
+            </button>
+            <button mat-icon-button matTooltip="Logout" (click)="logout()" routerLink="/" *ngIf="isAuthenticated()">
+                <mat-icon>exit_to_app</mat-icon>
+            </button>
+            <a mat-icon-button matTooltip="Github" href="http://github.com/DashboardHub/PipelineDashboard" target="_blank">
+                <mat-icon svgIcon="assets:github">aa</mat-icon>
+            </a>
+        </mat-toolbar>
+
+        <div class="content-container">
+            <router-outlet></router-outlet>
+        </div>
+
+        <mat-toolbar fxLayout="row" fxLayoutGap="10px">
+            <span class="mat-caption">Copyright &copy; 2018 <a href="https://www.dashboardhub.io" target="_blank" rel="noopener">DashboardHub</a>. All rights reserved <i>({{ version }})</i></span>
+            <span fxFlex></span>
+            <button mat-button (click)="showDoorbell()" color="accent">Feedback</button>
+            <mat-icon [style.width.px]="120" svgIcon="assets:dashboardhub"></mat-icon>
+        </mat-toolbar>
+    </mat-sidenav-content>
+</mat-sidenav-container>
