+++ web/src/app/app.component.ts
@@ -1,16 +1,68 @@
 import { Component } from '@angular/core';
 import { DomSanitizer } from '@angular/platform-browser';
 import { MatIconRegistry } from '@angular/material/icon';
+import { Navigation } from './shared/models/navigation.model';
+import { environment } from './../environments/environment';
+import { ProfileModel } from './shared/models/index.model';
+import { AuthenticationService } from './core/services/authentication.service';
 
 @Component({
     selector: 'dashboard-app',
-    template: '<router-outlet></router-outlet>'
+    templateUrl: './app.component.html',
+    styleUrls: ['./app.component.scss']
 })
 export class AppComponent {
-
+    public version: string;
+    public publicRoutes: Navigation[] = [
+        {
+            title: 'Homepage',
+            route: '/',
+            icon: 'home',
+        },
+    ];
+    public privateRoutes: Navigation[] = [
+        {
+            title: 'My Projects',
+            route: '/projects',
+            icon: 'group_work',
+        },
+        {
+            title: 'Add Project',
+            route: '/projects/create',
+            icon: 'playlist_add',
+        },
+        {
+            title: 'Profile',
+            route: '/profile',
+            icon: 'security',
+        },
+    ];
+    public generalRoutes: Navigation[] = [
+        {
+            title: 'Features',
+            route: '/features',
+            icon: 'lightbulb_outline'
+        },
+        {
+            title: 'Help',
+            route: '/help',
+            icon: 'live_help'
+        },
+        {
+            title: 'Terms & Conditions',
+            route: '/terms-and-conditions',
+            icon: 'copyright'
+        },
+        {
+            title: 'Privacy',
+            route: '/privacy',
+            icon: 'gavel'
+        }
+    ];
     constructor(
         private _iconRegistry: MatIconRegistry,
-        private _domSanitizer: DomSanitizer
+        private _domSanitizer: DomSanitizer,
+        private authService: AuthenticationService,
     ) {
         this._iconRegistry
             .addSvgIconInNamespace('assets', 'dashboardhub',
@@ -21,7 +73,7 @@ export class AppComponent {
             .addSvgIconInNamespace('assets', 'dashboardhub_white',
                 this._domSanitizer
                     .bypassSecurityTrustResourceUrl('https://raw.githubusercontent.com/DashboardHub/Assets/master/logo-horizontal-white.svg')
-                );
+            );
         this._iconRegistry
             .addSvgIconInNamespace('assets', 'dashboardhub_icon',
                 this._domSanitizer
@@ -32,5 +84,26 @@ export class AppComponent {
                 this._domSanitizer
                     .bypassSecurityTrustResourceUrl('https://raw.githubusercontent.com/DashboardHub/Assets/master//github.svg')
             );
+        this.version = environment.version;
+    }
+
+    public getProfile(): ProfileModel {
+        return this.authService.profile;
+    }
+
+    public isAuthenticated(): boolean {
+        return this.authService.isAuthenticated;
+    }
+
+    public login(): void {
+        this.authService.login();
+    }
+
+    public logout(): void {
+        this.authService.logout();
+    }
+
+    public showDoorbell(): void {
+        (<any>window).doorbell.show();
     }
 }
