+++ web/src/app/main/main-routing.module.ts
@@ -4,12 +4,6 @@ import { Routes, RouterModule } from '@angular/router';
 // Dashboard hub components
 import { MainComponent } from './main.component';
 import { HomepageComponent } from './components/homepage/homepage.component';
-import { CreateProjectComponent } from '../projects/create/create.component';
-import { ViewProjectComponent } from '../projects/view/view.component';
-import { ViewProjectResolver } from '../core/resolvers/view-project.resolver';
-import { EditProjectComponent } from '../projects/edit/edit.component';
-import { EditProjectResolver } from '../core/resolvers/edit-project.resolver';
-import { PrivateProjectsComponent } from '../projects/private/private.component';
 import { ProfileComponent } from './components/profile/profile.component';
 import { FeaturesComponent } from './components/features/features.component';
 import { HelpComponent } from './components/help/help.component';
@@ -26,30 +20,6 @@ const routes: Routes = [
             {
                 path: '',
                 component: HomepageComponent
-            },
-            {
-                path: 'projects/create',
-                pathMatch: 'full',
-                component: CreateProjectComponent,
-                canActivate: [AuthGuard]
-            },
-            {
-                path: 'project/:uid',
-                pathMatch: 'full',
-                component: ViewProjectComponent,
-                resolve: { project: ViewProjectResolver }
-            },
-            {
-                path: 'project/:uid/edit',
-                pathMatch: 'full',
-                component: EditProjectComponent,
-                resolve: { project: EditProjectResolver }
-            },
-            {
-                path: 'projects',
-                pathMatch: 'full',
-                component: PrivateProjectsComponent,
-                canActivate: [AuthGuard]
             },
             {
                 path: 'profile',
@@ -76,8 +46,7 @@ const routes: Routes = [
                 path: 'privacy',
                 pathMatch: 'full',
                 component: PrivacyComponent
-            },
-            { path: '**', redirectTo: '/' }
+            }
         ]
     }
 ];
