+++ web/src/app/projects/edit/edit.component.html
@@ -38,7 +38,7 @@
             <mat-divider></mat-divider>
             <mat-card-actions>
                 <div fxLayout="row" fxLayoutGap="20px" fxLayoutAlign="center center">
-                    <button mat-stroked-button type="button" [routerLink]="['/project', uid]">Cancel</button>
+                    <button mat-stroked-button type="button" (click)="onClickCancel()">Cancel</button>
                     <button mat-raised-button type="submit" color="accent" [disabled]="projectForm.invalid">Save</button>
                 </div>
             </mat-card-actions>
