+++ web/src/app/projects/view/view.component.html
@@ -7,7 +7,7 @@
         <span fxFlex></span>
         <span fxLayout="row" fxLayoutGap="20px" fxLayoutAlign="center center" *ngIf="isAdmin()">
             <button mat-raised-button type="submit" color="accent" (click)="addRepository()">Add Repository</button>
-            <button mat-raised-button type="submit" color="accent" [routerLink]="['/project', project.uid, 'edit']">Edit</button>
+            <button mat-raised-button type="submit" color="accent" [routerLink]="['edit']">Edit</button>
             <button mat-raised-button type="submit" color="warn" (click)="delete()">Delete</button>
         </span>
     </mat-toolbar>
