+++ web/src/app/shared/components/projects/projects.component.html
@@ -3,7 +3,7 @@
     <mat-list-item *ngFor="let project of projects" fxLayout='column'>
         <mat-icon mat-list-icon [matBadge]="project.repositories.length" matBadgeColor="accent" *ngIf="project.type === 'public'">lock_open</mat-icon>
         <mat-icon mat-list-icon [matBadge]="project.repositories.length" matBadgeColor="accent" *ngIf="project.type === 'private'">lock</mat-icon>
-        <h4 mat-line><a [routerLink]="['/project', project.uid]">{{ project.title }}</a></h4>
+        <h4 mat-line><a [routerLink]="['/projects', project.uid]">{{ project.title }}</a></h4>
         <p mat-line>{{ project.description }}</p>
         <span fxFlex></span>
         <p>{{ project.updatedOn.toDate() | timeAgo }}</p>
