+++ web/src/app/shared/components/projects/projects.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { ProjectModel } from '../shared/models/index.model';
+import { ProjectModel } from '../../models/index.model';
 
 @Component({
     selector: 'dashboard-projects',
