+++ web/src/app/shared/components/public/public.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
-import { ProjectService } from '../../core/services/project.service';
-import { ProjectModel } from '../../shared/models/index.model';
+import { ProjectService } from '../../../core/services/project.service';
+import { ProjectModel } from '../../models/index.model';
 import { Subscription } from 'rxjs';
 
 @Component({
